requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
numpy==1.14.5
tensorflow-gpu==1.10.0
pydot==1.4.1
matplotlib==2.1.0.
emnist
h5py==2.8.0
pandas==0.23.1
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
virtualenv==16.0.0